requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
jinja2==3.1.4
beautifulsoup4==4.12.3
groq==0.11.0
psycopg2-binary
